backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
feedparser==6.0.10
cohere==4.47
pinecone-client==3.0.2
python-dotenv==1.0.1
groq==0.4.2
pydantic==2.6.3
python-multipart==0.0.9
httpx==0.27.0
beautifulsoup4==4.12.3
